[CC-20] Extract DocumentResultGroup from homepage search results
Pulls the per-document result rendering (title/match-count header with
expand-collapse toggle, plus the ConfessionChapterResult/ConfessionTextResult
branching) out of pages/index.jsx's renderResults into its own component.
Pure refactor of the homepage - no behavior change - so the per-entry
pages landing in the next commit can render results identically instead
of reimplementing this logic.
